+++ changelog.md
@@ -4,6 +4,9 @@
 
 - **Fixed:**
   - Fix auto update progress dialog not updating title or progress state
+  - Datagrid: 
+    - Fix datagrid edit context item
+    - Make edit context item edit item clicked on instead of first selected item
 - **Changed:**
   - Update electron 21.4.2 to 21.4.3
   - Update mathjs 11.6.0 to 11.8.0